Rename sqlanalytics -> sql
