+++ sql/acceptance/sql_global_config_test.go
