+++ sql/resource_dashboard.go
@@ -1,11 +1,11 @@
-package sqlanalytics
+package sql
 
 import (
 	"context"
 	"fmt"
 
 	"github.com/databrickslabs/terraform-provider-databricks/common"
-	"github.com/databrickslabs/terraform-provider-databricks/sqlanalytics/api"
+	"github.com/databrickslabs/terraform-provider-databricks/sql/api"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
