+++ sql/resource_sql_endpoint.go
@@ -1,4 +1,4 @@
-package sqlanalytics
+package sql
 
 import (
 	"context"
