+++ sql/resource_sql_global_config.go
@@ -1,4 +1,4 @@
-package sqlanalytics
+package sql
 
 import (
 	"context"
