+++ sql/resource_sql_global_config_test.go
@@ -1,4 +1,4 @@
-package sqlanalytics
+package sql
 
 import (
 	"testing"
